Added the reshape function to pad inputs to the max patient shape for stacking
This was because the regular numpy resize rearranged pixels instead of padding them. Additionally, some variable names were changed and added a new parameter `max_patient_shape` so that padding is reliable (the filler value was x_train.min() because the input data was already locally preprocessed and the filler value was no longer `0`).