Huge changes to the transforms_generator generators
Small changes: removed the "_" from some methods to allow for inheritance.
Big: Made BaseTransformGenerator's data_gen method an abstract method. Then I made the Transformed2DGenerator and Transformed3DGenerator based on those. The Transformed3DGenerator is the basic one I had before. The Transformed2DGenerator is a bit different, where it only returns slices (and allows you the option to set the lower limit of the number of positive slices you want in a batch). (NOTE: Tests are not written yet.)